requirements.txt
notebook
matplotlib
numpy<2
openpyxl
pandas
scikit-learn
scipy
tqdm
umap-learn
torch
